% Auto-generated: do not edit by hand
\name{cascader}

\alias{cascader}

\title{Cascader component}

\description{
Cascader is a component that creates a dropdown select in form of cascade. It takes a property, `data` It renders the data and the value selected is returned by the property `value`
}

\usage{
cascader(id=NULL, data=NULL, value=NULL, height=NULL,
menuHeight=NULL, menuWidth=NULL, preventOverflow=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Unnamed list. Data that will be used in the component.}

\item{value}{Character. The value displayed capture by the component.}

\item{height}{Numeric. The height of the Dropdown.}

\item{menuHeight}{Numeric. The height of the menu.}

\item{menuWidth}{Numeric. The width of the menu.}

\item{preventOverflow}{Logical. The prevent the overflow  of the menu.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

